+++ src/main/kotlin/com/github/bobi/aemgroovyconsoleplugin/dsl/AemGdslProvider.kt
@@ -6,4 +6,4 @@ import org.jetbrains.plugins.groovy.dsl.GdslScriptProvider
  * User: Andrey Bardashevsky
  * Date/Time: 30.07.2022 23:15
  */
-class AemGDslProvider: GdslScriptProvider
+class AemGdslProvider: GdslScriptProvider