+++ src/main/kotlin/com/github/bobi/aemgroovyconsoleplugin/utils/AemFileTypeUtils.kt
@@ -1,6 +1,8 @@
 package com.github.bobi.aemgroovyconsoleplugin.utils
 
 import com.github.bobi.aemgroovyconsoleplugin.lang.AemConsoleScriptType
+import com.github.bobi.aemgroovyconsoleplugin.services.RootFoldersService
+import com.intellij.openapi.project.Project
 import com.intellij.openapi.vfs.VirtualFile
 import com.intellij.psi.PsiElement
 import com.intellij.psi.PsiFile
@@ -12,8 +14,9 @@ import com.intellij.psi.PsiFile
 @Suppress("unused")
 object AemFileTypeUtils {
 
-    fun VirtualFile.isAemFile(): Boolean {
-        return this.extension == AemConsoleScriptType.EXTENSION
+    fun VirtualFile.isAemFile(project: Project): Boolean {
+        return extension == AemConsoleScriptType.EXTENSION
+                || RootFoldersService.getInstance(project).isInsideAnyRoot(path)
     }
 
     fun PsiElement.isAemFile(): Boolean {
@@ -24,9 +27,13 @@ object AemFileTypeUtils {
         var virtualFile = this.virtualFile
 
         if (virtualFile == null) {
+            virtualFile = this.originalFile.virtualFile
+        }
+
+        if (virtualFile == null) {
             virtualFile = this.viewProvider.virtualFile
         }
 
-        return virtualFile.isAemFile()
+        return virtualFile.isAemFile(this.project)
     }
 }